wallet-ext: required unlocked account for pages
* disable access to pages that require unlocked accounts
* redirect to tokens page when account is locked
