+++ apps/wallet/src/ui/app/pages/home/nft-transfer/index.tsx
@@ -9,14 +9,16 @@ import Loading from '_components/loading';
 import { NFTDisplayCard } from '_components/nft-display';
 import Overlay from '_components/overlay';
 import { useOwnedNFT } from '_hooks';
+import { useUnlockedGuard } from '_src/ui/app/hooks/useUnlockedGuard';
 
 function NftTransferPage() {
 	const { nftId } = useParams();
 	const address = useActiveAddress();
 	// verify that the nft is owned by the user and is transferable
-	const { data: ownedNFT, isLoading } = useOwnedNFT(nftId || '', address);
+	const { data: ownedNFT, isLoading: isNftLoading } = useOwnedNFT(nftId || '', address);
 	const navigate = useNavigate();
-
+	const isGuardLoading = useUnlockedGuard();
+	const isLoading = isNftLoading || isGuardLoading;
 	return (
 		<Overlay showModal={true} title="Send NFT" closeOverlay={() => navigate('/nfts')}>
 			<div className="flex w-full flex-col h-full">
