+++ apps/wallet/src/ui/app/pages/home/transactions/index.tsx
@@ -9,6 +9,7 @@ import { QredoPendingTransactions } from './QredoPendingTransactions';
 import FiltersPortal from '_components/filters-tags';
 import { isQredoAccountSerializedUI } from '_src/background/accounts/QredoAccount';
 import { useActiveAccount } from '_src/ui/app/hooks/useActiveAccount';
+import { useUnlockedGuard } from '_src/ui/app/hooks/useUnlockedGuard';
 import PageTitle from '_src/ui/app/shared/PageTitle';
 
 function TransactionBlocksPage() {
@@ -16,6 +17,9 @@ function TransactionBlocksPage() {
 	const isQredoAccount = !!(activeAccount && isQredoAccountSerializedUI(activeAccount));
 	const { status } = useParams();
 	const isPendingTransactions = status === 'pending';
+	if (useUnlockedGuard()) {
+		return null;
+	}
 	if (activeAccount && !isQredoAccount && isPendingTransactions) {
 		return <Navigate to="/transactions" replace />;
 	}
